0818_マイグレーションファイルの修正 #265

+++ db/migrate/20220717162248_create_members.rb
@@ -1,8 +1,8 @@
 class CreateMembers < ActiveRecord::Migration[6.1]
   def change
     create_table :members do |t|
-      t.integer :team_id
-      t.integer :user_id
+      t.bigint :team_id
+      t.bigint :user_id
 
       t.timestamps
     end
