+++ db/migrate/20220814052629_create_ticket_categorle.rb
@@ -1,8 +1,8 @@
 class CreateTicketCategorle < ActiveRecord::Migration[6.1]
   def change
     create_table :ticket_categorles do |t|
-      t.integer :ticket_id
-      t.integer :categorle_id
+      t.bigint :ticket_id
+      t.bigint :categorle_id
 
       t.timestamps
     end
